project/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.3.0",
    "flet",
    "mysql-connector-python>=9.3.0",
]
